@@ -5,7 +5,7 @@ import lombok.experimental.Accessors;
 import org.springframework.data.annotation.CreatedDate;
 import org.springframework.data.jpa.domain.support.AuditingEntityListener;
 
-import javax.persistence.*;
+import jakarta.persistence.*;
 import java.io.Serializable;
 import java.time.LocalDateTime;
 import java.util.List;
@@ -42,8 +42,6 @@ public class User implements Serializable {
 
 
     @OneToOne(fetch = FetchType.LAZY, mappedBy = "user")
-    @Setter(AccessLevel.NONE)
-    @Getter(AccessLevel.NONE)
     private Purse purse;
 
 
